ref(ca2): changing type hint of lat and lng of Location

+++ CA2/darooghe/domain/entity/location.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
-
+from decimal import Decimal
 
 @dataclass
 class Location:
-    lat: float
-    lng: float
+    lat: Decimal
+    lng: Decimal
